pyproject.toml
[tool.poetry]
name = "Moradia IFPE"
version = "0.1.0"
description = "Sistema de Gestão do Programa de Moradia do IFPE"
authors = ["Ismael Benjamim <ismaelbenjamim15@hotmail.com>", "Marco Antônio <maamo@cin.ufpe.br>",
    "Barbara Vaz <bvf@cin.ufpe.br>", "Liliane Carla <lco2@cin.ufpe.br>", "Vinícius Oliveira <vmo2@cin.ufpe.br>"]
license = "CC BY-NC-ND"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2"
python-decouple = "^3.4"
dj-database-url = "^0.5.0"
gunicorn = "^20.1.0"
djangorestframework = "^3.13.1"
django-rest-auth = "^0.9.5"
drf-yasg = "^1.20.0"
uuid = "^1.30"
psycopg2-binary = "^2.9.3"
django-heroku = "^0.3.1"
dj-static = "^0.0.6"
django-cors-headers = "^3.13.0"
Serializer = "^0.2.1"
pandas = "^1.4.3"
xlrd = "^2.0.1"
django-extensions = "^3.2.0"
django-webpack-loader = "^1.6.0"
Pillow = "^9.2.0"


[tool.poetry.dev-dependencies]
pytest-django = "^4.2.0"
flake8 = "^3.9.1"
pytest-cov = "^2.11.1"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "moradia_ifpe.settings"
python_files = [
    "test*.py",
     "*tests.py",
]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "moradia_ifpe/core/",
]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
